Fixing code review
