+++ classes/privacy/provider.php
@@ -21,7 +21,6 @@
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 namespace qtype_crossword\privacy;
-defined('MOODLE_INTERNAL') || die();
 /**
  * Privacy Subsystem for qtype_crossword implementing null_provider.
  *
