+++ db/install.xml
@@ -27,11 +27,11 @@
       <FIELDS>
         <FIELD NAME="id" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="true"/>
         <FIELD NAME="questionid" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Redundant, because of the answer field. Foreign key references question.id."/>
-        <FIELD NAME="answer" TYPE="char" LENGTH="200" NOTNULL="true" SEQUENCE="false" COMMENT="The answer text."/>
         <FIELD NAME="clue" TYPE="char" LENGTH="200" NOTNULL="true" SEQUENCE="false" COMMENT="The clue text."/>
+        <FIELD NAME="answer" TYPE="char" LENGTH="200" NOTNULL="true" SEQUENCE="false" COMMENT="The answer text."/>
+        <FIELD NAME="startcolumn" TYPE="int" LENGTH="2" NOTNULL="false" SEQUENCE="false" COMMENT="The vertical index."/>
+        <FIELD NAME="startrow" TYPE="int" LENGTH="2" NOTNULL="true" SEQUENCE="false" COMMENT="Horizontal index."/>
         <FIELD NAME="orientation" TYPE="int" LENGTH="1" NOTNULL="true" DEFAULT="0" SEQUENCE="false" COMMENT="The orientation option."/>
-        <FIELD NAME="rowindex" TYPE="int" LENGTH="2" NOTNULL="true" SEQUENCE="false" COMMENT="Horizontal index."/>
-        <FIELD NAME="columnindex" TYPE="int" LENGTH="2" NOTNULL="false" SEQUENCE="false" COMMENT="The vertical index."/>
       </FIELDS>
       <KEYS>
         <KEY NAME="primary" TYPE="primary" FIELDS="id"/>
