+++ styles.css
@@ -7,7 +7,7 @@
 }
 
 #page-question-type-crossword .mform .form-group.has-danger .col-md-9[data-fieldtype="group"] {
-    border: none!important;
+    border: none !important;
 }
 
 #page-question-type-crossword .crossword-contain {
@@ -31,13 +31,12 @@
 }
 
 #page-question-type-crossword fieldset#id_wordhdr .d-flex .form-group:last-child,
-.path-question-bank-previewquestion .contain-crossword .wrap-clue img.icon,
-.path-mod-quiz .contain-crossword .wrap-clue img.icon {
+.qtype_crossword-grid-wrapper .wrap-clue img.icon {
     margin-right: 0;
 }
 
-#page-question-type-crossword fieldset#id_wordhdr .fcontainer input[name^="answer"],
-.path-question-bank-previewquestion .wrap-crossword #crossword form .crossword .formulation input[type="text"]{
+#page-question-type-crossword fieldset#id_words .fcontainer input[name^="answer"],
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword form .crossword .formulation input[type="text"]{
     text-transform: uppercase;
 }
 
@@ -77,7 +76,6 @@
 
 #page-question-type-crossword .grid-square .word-label {
     position: absolute;
-    z-index: 1;
     top: 2px;
     left: 2px;
     line-height: 1;
@@ -86,72 +84,52 @@
     pointer-events: none;
 }
 
-#page-mod-quiz-reviewquestion section#region-main .contain-crossword div div {
+#page-mod-quiz-reviewquestion section#region-main .qtype_crossword-grid-wrapper div div {
     margin: 0;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword,
-.path-question .wrap-crossword #crossword,
-.path-mod-quiz .wrap-crossword #crossword {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword {
     position: relative;
     max-width: 700px;
     background: #000000;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-cell,
-.path-question .wrap-crossword #crossword .crossword-cell,
-.path-mod-quiz .wrap-crossword #crossword .crossword-cell {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-cell {
     cursor: pointer;
     fill: #ffffff;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-grid-background,
-.path-question .wrap-crossword #crossword .crossword-grid-background,
-.path-mod-quiz .wrap-crossword #crossword .crossword-grid-background {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-grid-background {
     fill: #000000;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-cell-highlighted,
-.path-question .wrap-crossword #crossword .crossword-cell-highlighted,
-.path-mod-quiz .wrap-crossword #crossword .crossword-cell-highlighted {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-cell-highlighted {
     fill: #ffe500;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-cell-focussed,
-.crossword-grid-focussed .crossword-cell-focussed,
-.path-question .wrap-crossword #crossword .crossword-cell-focussed,
-.path-mod-quiz .wrap-crossword #crossword .crossword-cell-focussed {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-cell-focussed {
     fill: #fff7b2;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-cell-number,
-.path-question .wrap-crossword #crossword .crossword-cell-number,
-.path-mod-quiz .wrap-crossword #crossword .crossword-cell-number {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-cell-number {
     font-size: 0.625rem;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-hidden-input-wrapper,
-.path-question .wrap-crossword #crossword .crossword-hidden-input-wrapper,
-.path-mod-quiz .wrap-crossword #crossword .crossword-hidden-input-wrapper {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-hidden-input-wrapper {
     display: none;
     position: absolute;
-    z-index: 1;
     top: 0;
     left: 0;
     width: 6.6%;
     height: 6.6%;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-cell-text,
-.path-question .wrap-crossword #crossword .crossword-cell-text,
-.path-mod-quiz .wrap-crossword #crossword .crossword-cell-text {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-cell-text {
     text-transform: uppercase;
     font-size: 0.65rem;
 }
 
-.path-question-bank-previewquestion .wrap-crossword #crossword .crossword-hidden-input-wrapper input,
-.path-question .wrap-crossword #crossword .crossword-hidden-input-wrapper input,
-.path-mod-quiz .wrap-crossword #crossword .crossword-hidden-input-wrapper input {
+.qtype_crossword-grid-wrapper .wrap-crossword #crossword .crossword-hidden-input-wrapper input {
     border: 0;
     padding: 0;
     width: 100%;
@@ -166,9 +144,7 @@
     caret-color: transparent;
 }
 
-.path-question-bank-previewquestion .contain-crossword .wrap-clue .clue-id,
-.path-question .contain-crossword .wrap-clue .clue-id,
-.path-mod-quiz .contain-crossword .wrap-clue .clue-id {
+.qtype_crossword-grid-wrapper .wrap-clue .clue-id {
     border-radius: 50%;
     flex-shrink: 0;
     margin-right: 0.2em;
@@ -177,60 +153,38 @@
     height: 1.75em;
 }
 
-.path-question-bank-previewquestion .contain-crossword .wrap-clue .clue-content,
-.path-question .contain-crossword .wrap-clue .clue-content,
-.path-mod-quiz .contain-crossword .wrap-clue .clue-content {
+.qtype_crossword-grid-wrapper .wrap-clue .clue-content {
     word-break: break-word;
 }
 
-.path-question-bank-previewquestion .contain-crossword .wrap-clue.focus .clue-id,
-.path-question .contain-crossword .wrap-clue.focus .clue-id,
-.path-mod-quiz .contain-crossword .wrap-clue.focus .clue-id {
+.qtype_crossword-grid-wrapper .wrap-clue.focus .clue-id {
     background: #ffe500;
 }
 
-.path-question-bank-previewquestion .contain-crossword .wrap-clue input,
-.path-question .contain-crossword .wrap-clue input,
-.path-mod-quiz .contain-crossword .wrap-clue input {
+.qtype_crossword-grid-wrapper .wrap-clue input[type="text"] {
     width: calc(100% - 3em);
     letter-spacing: 0.3em;
     text-transform: uppercase;
 }
 
-.path-question-bank-previewquestion .wrap-crossword,
-.path-question .wrap-crossword,
-.path-mod-quiz .wrap-crossword {
-    position: relative;
-}
-
-.path-question-bank-previewquestion .wrap-crossword .sticky-clue,
-.path-question .wrap-crossword .sticky-clue,
-.path-mod-quiz .wrap-crossword .sticky-clue {
-    position: absolute;
+.qtype_crossword-grid-wrapper .wrap-crossword .sticky-clue {
     top: 0;
     left: 0;
 }
 
 /* Small devices (landscape phones, less than 768px). */
 @media (max-width: 767.98px) {
-    .path-question-bank-previewquestion .contain-crossword .clue-id > span,
-    .path-question .contain-crossword .clue-id > span,
-    .path-mod-quiz .contain-crossword .clue-id > span {
+    .qtype_crossword-grid-wrapper .clue-id > span {
         font-weight: bold;
         font-style: normal;
     }
 
-    .path-question-bank-previewquestion .wrap-crossword #crossword .orientation,
-    .path-question .wrap-crossword #crossword .orientation,
-    .path-mod-quiz .wrap-crossword #crossword .orientation {
+    .qtype_crossword-grid-wrapper .wrap-crossword #crossword .orientation {
         margin-bottom: 0;
     }
 
-    .path-question-bank-previewquestion .wrap-crossword,
-    .path-question .wrap-crossword,
-    .path-mod-quiz .wrap-crossword {
+    .qtype_crossword-grid-wrapper .wrap-crossword {
         position: relative;
-        padding-top: 2.5rem;
         padding-bottom: 1.25rem;
     }
 }
@@ -247,9 +201,7 @@
 }
 
 @media print {
-    .path-question-bank-previewquestion .contain-crossword #crossword svg g rect,
-    .path-question .contain-crossword #crossword svg g rect,
-    .path-mod-quiz .contain-crossword #crossword svg g rect {
+     .qtype_crossword-grid-wrapper #crossword svg g rect {
         stroke: #000000;
         stroke-width: 0.5px;
     }
